+++ plugin/src/main/kotlin/com/linecorp/intellij/plugins/armeria/explorer/ArmeriaRouteCollector.kt
@@ -139,6 +139,7 @@ object ArmeriaRouteCollector {
                 routeMatch = RouteMatch.ANNOTATED_HTTP,
                 decorators = methodDecorators.distinct(),
                 exceptionHandlers = methodExceptionHandlers.distinct(),
+                timeoutHints = ArmeriaTimeoutSupport.collectTimeoutHints(method),
             )
         }
     }